Rip apart graph test
